+++ NEWS.md
@@ -23,6 +23,7 @@ Enhancements:
   * messages and warnings are now printed using message() and warning()
   * Stop if any of the input files has duplicated column names
   * Warn user if success.arrays is not defined in migration()
+  * 'Detections' column in the object *status.df* has been split into 'Valid.detections' and 'All.detections'.
 
 ## actel 0.1.1